Allow SVG::Document to write out the SVG to a location
I went looking around for how to do this but didn't come up with a
great solution. The weird part to me is the `String.to_slice` method.
IO.write only takes a ByteSlice, not a string and this is the only way
to get one from a string. I also looked at another way of doing it by
copying from one IO to another, but I'm not sure how the data got into
the IO in the first place...