Add StartPointFinder
This simply finds the right-most point of the bottom-most row of a set
of points. In a scattering of points, this would most likely find simply
the bottom point. But if you have a grid of points, there may be
mutliple points that all have the same low Y coordinate.