Add Tracer class with angle_between
This will be used for getting the angle of point2 as if point1 was the
origin. I'll still have to shift the angle I get back by the previous
angle, but that's not for this method to do.

Also, CrystalEdge uses atan2 to calculate angle, which gives me the
angle from -π to π. However, I need to find the smallest angle from the
original point. An angle in the 0 - -π range is numerically smaller but
functionally bigger, so I needed to convert the angles from a 0 - 2π
range.