Add previous_point method to tracer
This may not be the best way to do this. This method is supporting the
algorithms need to know the angle of the vector created by distance
between the last two vectors. So, if you're moving from the tip of
Vector A to Vector B, what angle was that?

If you only have one point, like at the start of the algorithm, we want
a 0 degree angle. To get that, I have this method returning that initial
point. Then later, when comparing the angle between these "two" points,
it will return 0.

However, previous_point doesn't seem to really represent the concept
very well since sometimes it won't return the previous point but the
current one, like at the start of the algo. I suspect I'm going down the
wrong path a bit here.

I'm keeping it for now because I'd like to have something working that I
can refactor as opposed to trying to get it perfect on my first go.