NearestNeighbor finder
This finds the nearest N points to an origin point based on the
magnitude of the vector between two points. The one catch is, if there
is a tie for "nearest" that would return more than N points, it will
return them.

So there might be 6 points with the following magnitudes:

1, 3, 4, 4, 5, 8

If you ask for the 3 nearest points, it will return the points with 1,
3, 4 magnitudes, including both 4 magnitude points, which is 4 points
total.