Add SmallestAngle class
This class will return the point that is the smallest clockwise angle
from the previous angle and new origin point. I'll get more into that
when I write the code, but for now, all I need this class to do is
adjust the initial points so their origin is some arbitrary point. This
is so the points will be in the right place so that I can easlily do
calculations around the difference in angle between two points.