Add angle_to method to Vector2 class
This does will do a lot of my biasing work for me. It's job is to return
the angle of a vector between the two vectors, as if the tip of the
caller vector was pointing at the passed in vector.

By subtracting the caller from the passed in vector, we can get this new
vector that points from one to the other. Then we can simply get its
angle.