Find the point which has the smallest angle from last axis
Imagine a bunch of points. Pick some point to start with, call it
"origin" and select a few points nearby. Now pick the one that has the
smallest angle counter-clockwise around starting at 0 degrees, the
X-axis.

Let's say we pick a point that has an angle of 30 degrees. We use that
new point as the new origin. Then we do the same thing; pick some
nearby points, find the smallest counter-clockwise angle. Except,
instead of starting at 0 degrees, we use the inverse of the angle we
traveled to get to this point.

So we got here by going 30 degrees, so we'll do 30 - 180 degrees = -150
or 210 degrees and use that as our starting point for
counter-clockwise. Then rinse and repeat, new origin, new nearby
points, new angle, find the smallest counter-clockwise angle.

That's this algorithm in a nutshell.

This commit adds the different angle part, subtracting the angle of the
nearby points by whatever the offset angle is passed in.