Add GridNearestNeighbors class
This finds all points that are one unit away from an origin point.

Imagine a tic-tac-toe board. Put an X in the center to represent the
origin point. This finder will find all points on the outer edges of the
board. Diagonal points count as 1 unit away as well as points that are
one unit away on the X and Y axis.