Ensure the last direction traveled is the last to be considered
When looking for the point with the smallest angle from the last
direction travled, I sort the points from smallest to largest angle.

The angle is shifted by the last direction traveled. If the current
point was found by moving 45deg, SmallestAngle will shift 45deg to be
0deg, so that 46deg is a small angle difference instead of a larger one.

The problem is, the smallest angle is 0deg but that's back the
direction the we came! To make it the "last" angle considered, the
angle is shifted by 0.00001 radians. That way 0 degrees becomes
359.9999 and therefore the largest angle and the last one considered.