Refactor SmallestAngle
This problem here was that I was further offsetting the offset_angle
because if I didn't, it would always pick the point that I came from.

Because the algorithm picks the point with the smallest clockwise angle
from the last point, the last point is always at 0 radians/degrees.
Zero is usually considered pretty small.

Previously I was offsetting the angle by a very small amount so it
would list the point it came from as 359.9999 degrees instead of 0. It
worked, but it was gross.

Now I remove the point at 0 degrees because I never want that point and
then sort the rest of the points by smallest angle.

I also fixed a typo in the tests.