Add first version of point grouper
This will take points that are adjacent (one step in a cardinal or
diagonal direction) and put them together in an array. If a point is
separated, it and points adjacent to it will be in a different array.
The result is a 2-dimensional array of points. These arrays would then
get funneled into the Tracer class to just get their outlines.