HeightPoint
This type is here because the Vector2 class doesn't store a height
parameter and I didn't want to extend it in that way. So I wrapped it
in this class that has a height and point attribute.

I need the height because I'm about to sort points by their height
values.