Working LevelSplitter
This takes a list of points (HeightPoints) and sorts them into groups
by their height. It's logic is:

* Figure out the heights based on a start and step parameter
* Create ranges from those steps
* Select points that are covered by those ranges

I'm using Int32 for the start and step parameters because splitting up
floats into ranges isn't yet well supported in Crystal:
https://github.com/crystal-lang/crystal/issues/3021

I use an exclusive range for everything but the last range. This is so
all points are covered. Let's say we have points with heights ranging
from 0 to 10. The full range is 0 - 10 and we'll do steps of 5:

0...5 => 0, 1, 2, 3, 4
5...10 => 5, 6, 7, 8, 9

The points with a height of 10 would be missed. I could make the max 1
higher, but this results in an extra range that's just the top-most
point and probably wouldn't look great in a contour map.

So instead, I just make the last range an inclusive range:

5, 6, 7, 8, 9, 10

Visually, this will look just fine when drawing contours.