Refactor LevelSplitter
This does essentially the same thing, but with a little less control.
You don't get to specify where it starts collecting data, as it will
always start from the lowest point. This is to ensure that it will
always catch all the point and won't leave any hanging which would
cause an infinite loop here (while !points.empty?)

What this does it it:
* chooses a height level
* moves all the points that are below but not equal to that level into
an array
* puts that array in the larger array of all groups
* increases the height level

This goes until it has no more points left to find.